+++ lib/screens/settings_screen.dart
@@ -2,6 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
+import '../helper/i18n_helper.dart';
 import '../provider/settings_provider.dart';
 import '../widgets/app_drawer.dart';
 
@@ -13,34 +14,49 @@ class SettingsScreen extends StatefulWidget {
 }
 
 class _SettingsScreenState extends State<SettingsScreen> {
-  static const _startViewValues = {
-    StartView.AllPasswords: ["All Passwords", Icons.description],
-    StartView.Folders: ["Folders", Icons.folder_open],
-    StartView.Favorites: ["Favorites", Icons.star],
-  };
-
-  final _startViewMenuItems = _startViewValues.keys
-      .map(
-        (key) => DropdownMenuItem(
-          value: key,
-          child: Row(
-            children: [
-              Icon(_startViewValues[key][1]),
-              const SizedBox(
-                width: 5,
-              ),
-              Text(_startViewValues[key][0]),
-            ],
-          ),
-        ),
-      )
-      .toList();
-
   @override
   Widget build(BuildContext context) {
+    final _startViewValues = {
+      StartView.AllPasswords: [
+        tl(context, "general.all_passwords"),
+        Icons.description,
+      ],
+      StartView.Folders: [
+        tl(
+          context,
+          "general.folders",
+        ),
+        Icons.folder_open
+      ],
+      StartView.Favorites: [
+        tl(
+          context,
+          "general.favorites",
+        ),
+        Icons.star
+      ],
+    };
+
+    final _startViewMenuItems = _startViewValues.keys
+        .map(
+          (key) => DropdownMenuItem(
+            value: key,
+            child: Row(
+              children: [
+                Icon(_startViewValues[key][1]),
+                const SizedBox(
+                  width: 5,
+                ),
+                Text(_startViewValues[key][0]),
+              ],
+            ),
+          ),
+        )
+        .toList();
+
     return Scaffold(
       appBar: AppBar(
-        title: const Text('Settings'),
+        title: Text(tl(context, 'general.settings')),
       ),
       drawer: AppDrawer(),
       body: Padding(
@@ -51,16 +67,15 @@ class _SettingsScreenState extends State<SettingsScreen> {
             Row(
               mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: [
-                const Text(
-                  'Start view',
+                Text(
+                  tl(context, 'settings.start_view'),
                   style: TextStyle(fontSize: 16),
                 ),
                 Consumer<SettingsProvider>(
                   builder: (context, settings, child) => DropdownButton(
                     value: settings.startView,
                     onChanged: (value) {
-                      Provider.of<SettingsProvider>(context, listen: false)
-                          .startView = value;
+                      settings.startView = value;
                     },
                     items: _startViewMenuItems,
                   ),
@@ -73,18 +88,59 @@ class _SettingsScreenState extends State<SettingsScreen> {
             Row(
               mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: [
-                const Text(
-                  'Use biometric auth',
+                Text(
+                  tl(context, 'settings.biometric_auth'),
                   style: TextStyle(fontSize: 16),
                 ),
                 Consumer<SettingsProvider>(
                   builder: (context, settings, child) => Checkbox(
-                    value: Provider.of<SettingsProvider>(context, listen: false)
-                        .useBiometricAuth,
+                    value: settings.useBiometricAuth,
                     onChanged: (value) {
-                      Provider.of<SettingsProvider>(context, listen: false)
-                          .useBiometricAuth = value;
+                      settings.useBiometricAuth = value;
                     },
+                  ),
+                ),
+              ],
+            ),
+            const SizedBox(
+              height: 10,
+            ),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                Text(
+                  tl(context, 'settings.passwort_strength'),
+                  style: TextStyle(fontSize: 16),
+                ),
+                Expanded(
+                  child: Consumer<SettingsProvider>(
+                    builder: (context, settings, child) => Slider(
+                      value: settings.passwordStrength.toDouble(),
+                      min: 5,
+                      max: 30,
+                      divisions: 25,
+                      label: settings.passwordStrength.toString(),
+                      inactiveColor: settings.passwordStrength < 8
+                          ? Colors.red
+                          : settings.passwordStrength < 13
+                              ? Colors.orange
+                              : settings.passwordStrength < 20
+                                  ? Colors.yellow
+                                  : Colors.green,
+                      activeColor: settings.passwordStrength < 8
+                          ? Colors.red
+                          : settings.passwordStrength < 13
+                              ? Colors.orange
+                              : settings.passwordStrength < 20
+                                  ? Colors.yellow
+                                  : Colors.green,
+                      onChanged: (value) {
+                        settings.passwordStrengthNoWrite = value.round();
+                      },
+                      onChangeEnd: (value) {
+                        settings.passwordStrength = value.round();
+                      },
+                    ),
                   ),
                 ),
               ],