+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.9.4
+- Add yes/no dialog for logout
+- Add settings for the password strength
+- Add translations: german
+
 ## 0.9.3
 - Add settings screen:
     - Set start view: All Passwords, Folders or Favorites